requirements.txt
flask==2.3.3
google-generativeai==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pydantic==2.0.0
pytest==7.4.0
typing-extensions==4.8.0
